move %-format rewrite to a plugin
