+++ tests/identity_equality_test.py
@@ -1,6 +1,6 @@
 import pytest
 
-from pyupgrade._main import _fix_py2_compatible
+from pyupgrade._main import _fix_plugins
 
 
 @pytest.mark.parametrize(
@@ -20,7 +20,7 @@ from pyupgrade._main import _fix_py2_compatible
     ),
 )
 def test_fix_is_compare_to_literal_noop(s):
-    assert _fix_py2_compatible(s, (2, 7)) == s
+    assert _fix_plugins(s, min_version=(2, 7), keep_percent_format=False) == s
 
 
 @pytest.mark.parametrize(
@@ -61,5 +61,5 @@ def test_fix_is_compare_to_literal_noop(s):
     ),
 )
 def test_fix_is_compare_to_literal(s, expected):
-    ret = _fix_py2_compatible(s, (2, 7))
+    ret = _fix_plugins(s, min_version=(2, 7), keep_percent_format=False)
     assert ret == expected
