+++ tests/set_literals_test.py
@@ -2,7 +2,7 @@ import sys
 
 import pytest
 
-from pyupgrade._main import _fix_py2_compatible
+from pyupgrade._main import _fix_plugins
 
 
 @pytest.mark.parametrize(
@@ -19,7 +19,7 @@ from pyupgrade._main import _fix_py2_compatible
     ),
 )
 def test_fix_sets_noop(s):
-    assert _fix_py2_compatible(s, (2, 7)) == s
+    assert _fix_plugins(s, min_version=(2, 7), keep_percent_format=False) == s
 
 
 @pytest.mark.parametrize(
@@ -94,7 +94,7 @@ def test_fix_sets_noop(s):
     ),
 )
 def test_sets(s, expected):
-    ret = _fix_py2_compatible(s, (2, 7))
+    ret = _fix_plugins(s, min_version=(2, 7), keep_percent_format=False)
     assert ret == expected
 
 
@@ -114,5 +114,5 @@ def test_sets(s, expected):
     ),
 )
 def test_sets_generators_trailing_commas(s, expected):
-    ret = _fix_py2_compatible(s, (2, 7))
+    ret = _fix_plugins(s, min_version=(2, 7), keep_percent_format=False)
     assert ret == expected
